Rename density.hpp to base.hpp
